feat(domain): Festival, Artist, School model 추가
