+++ android/festago/domain/src/main/java/com/festago/festago/domain/model/artist/Artist.kt
@@ -1,0 +1,7 @@
+package com.festago.festago.domain.model.artist
+
+data class Artist(
+    val id: Long,
+    val name: String,
+    val imageUrl: String,
+)
