+++ android/festago/domain/src/main/java/com/festago/festago/domain/model/school/School.kt
@@ -1,0 +1,7 @@
+package com.festago.festago.domain.model.school
+
+data class School(
+    val id: Long,
+    val name: String,
+    val imageUrl: String,
+)
